fix(GRO-2876): expenditure[].balanceDeclared property must be Integer for the Lending API clients

+++ src/main/java/com/selina/lending/api/dto/common/ExpenditureDto.java
@@ -31,7 +31,7 @@ public class ExpenditureDto {
     @Schema(implementation = Frequency.class)
     @EnumValue(enumClass = Frequency.class)
     String frequency;
-    Double balanceDeclared;
+    Integer balanceDeclared;
 
     @NotNull
     Double amountDeclared;
